# internal/handler/v1/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a swagger specification for a Duel Duck back-end.",
        "title": "Duel Duck API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/stats/user": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves the referral statistics (such as total referrals, income, etc.) for the currently authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stats"
                ],
                "summary": "Get referral statistics for the authenticated user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Referral statistics retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/model.UserReferralStats"
                        }
                    },
                    "401": {
                        "description": "Authentication required or invalid token",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    },
                    "500": {
                        "description": "Internal server error during referral stats retrieval",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    }
                }
            }
        },
...
        "/user": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves the profile information of the currently authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get current authenticated user profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User profile retrieved successfully.",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    },
                    "401": {
                        "description": "Authentication required or invalid token",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    },
                    "500": {
                        "description": "Internal server error during user retrieval",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the profile information of the currently authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Edit current authenticated user profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "User profile update payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserEditReq"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "User profile updated successfully"
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    },
                    "401": {
                        "description": "Authentication required or invalid token",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    },
                    "500": {
                        "description": "Internal server error during user update",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    }
                }
            }
        },
        "/user/profile-picture": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Uploads and sets a new profile picture for the currently authenticated user. The image file must be provided as multipart/form-data under the \"image\" field.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Set profile picture for the authenticated user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Profile picture image file",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profile picture updated successfully, returns the new image URL",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "image_url": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request data or missing image file",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    },
                    "401": {
                        "description": "Authentication required or invalid token",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    },
                    "500": {
                        "description": "Internal server error during profile picture update",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    }
                }
            }
        },
        "/user/username": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the username for the currently authenticated user. The new username must be valid and not already taken.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Change the username of the authenticated user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Username change payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UsernameChange"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Username changed successfully"
                    },
                    "400": {
                        "description": "Invalid request parameters or invalid username",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    },
                    "401": {
                        "description": "Authentication required or invalid token",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    },
                    "409": {
                        "description": "This username is already taken",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    },
                    "500": {
                        "description": "Internal server error during username change",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    }
                }
            }
        },
        "/user/username/{username}": {
            "get": {
                "description": "Retrieves user information by the provided username. This endpoint allows public access to view user profiles without authentication.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Find user by username",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username to search for",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User information retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    },
                    "400": {
                        "description": "Invalid username format",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/apperrors.ErrorPublic"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apperrors.ErrorPublic": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "model.User": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "integer"
                },
                "bg_url": {
                    "type": "string"
                },
                "bio": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "current_xp": {
                    "type": "integer"
                },
                "daily_reward_streak": {
                    "type": "integer"
                },
                "discord_url": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "instagram_url": {
                    "type": "string"
                },
                "is_premium": {
                    "type": "boolean"
                },
                "last_completed_streak": {
                    "type": "string"
                },
                "level": {
                    "type": "integer"
                },
                "public_address": {
                    "type": "string"
                },
                "referral_token": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/mtype.Role"
                },
                "sol_autoswap": {
                    "type": "boolean"
                },
                "telegram_id": {
                    "type": "string"
                },
                "telegram_url": {
                    "type": "string"
                },
                "two_fa": {
                    "type": "boolean"
                },
                "two_fa_secret": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "usdc_autoswap": {
                    "type": "boolean"
                },
                "username": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                },
                "x_url": {
                    "type": "string"
                },
                "youtube_url": {
                    "type": "string"
                }
            }
        },
        "model.UserEditReq": {
            "type": "object",
            "properties": {
                "discord_url": {
                    "type": "string"
                },
                "instagram_url": {
                    "type": "string"
                },
                "sol_autoswap": {
                    "type": "boolean"
                },
                "telegram_url": {
                    "type": "string"
                },
                "usdc_autoswap": {
                    "type": "boolean"
                },
                "website": {
                    "type": "string"
                },
                "x_url": {
                    "type": "string"
                },
                "youtube_url": {
                    "type": "string"
                }
            }
        },
...
        "model.UserReferralStats": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "referral_count": {
                    "type": "integer"
                },
                "referral_income_dp": {
                    "type": "integer"
                },
                "referral_income_usdc": {
                    "type": "number"
                },
                "referral_token": {
                    "type": "string"
                }
            }
        },
        "model.UsernameChange": {
            "type": "object",
            "properties": {
                "username": {
                    "type": "string"
                }
            }
        },
...
        "mtype.Role": {
            "type": "integer",
            "enum": [
                0,
                1
            ],
            "x-enum-varnames": [
                "UserRole",
                "AdminRole"
            ]
        },
...
    }
}